requirements.txt
Django>=4.2
djangorestframework
channels
channels_redis
drf-yasg
psycopg2-binary>=2.9.9
python-decouple